// src/app/erp/dashboard/component/parametros/state/interface/clase.interface.ts
export interface Clase {
    clasTipId: number;
    empId: number;
    clasTipDes: string;
    clasTipCod: string;
    clasTip: string;
    created_at: string;
    updated_at: string;
    // Agrega más propiedades según necesites
}

export const CLASE_KEYS = [
    "clasTipId",
    "empId",
    "clasTipDes",
    "clasTip",
    "created_at",
    "updated_at"

    // Agrega más keys según necesites
];